Only allow swipe to dismiss on pager's first page
